added input

+++ Jenkinsfile
@@ -49,12 +49,10 @@ pipeline {
                         message: 'What is the reg value?', 
                         parameters: [
                             [$class: 'ChoiceParameterDefinition', 
-                                choices: 'Choice 1\nChoice 2\nChoice 3', 
+                                choices: 'yes\nno', 
                                 name: 'input', 
                                 description: 'A select box option']
                         ])
-
-                    echo "Reg is ${reg}"
                 }
                 script {
 
@@ -66,21 +64,26 @@ pipeline {
                sh 'echo "Deploying..."'
                withCredentials([sshUserPrivateKey(credentialsId: 'ww-prod-cred', keyFileVariable: 'SSH_KEY')]) {
                 script {
-                    try {
+                    if(reg = 1) {
+                        try {
                         sh '''
                              ssh -o StrictHostKeyChecking=no -i $SSH_KEY ec2-user@ec2-13-58-233-86.us-east-2.compute.amazonaws.com 'rm -r WWBUILD'
                         '''
                     } catch(Exception e) {
 
                     }
-                }
-                sh '''
-                    ssh -o StrictHostKeyChecking=no -i $SSH_KEY ec2-user@ec2-13-58-233-86.us-east-2.compute.amazonaws.com 'mkdir WWBUILD'
-                    scp -o StrictHostKeyChecking=no -i $SSH_KEY -r backend ec2-user@ec2-13-58-233-86.us-east-2.compute.amazonaws.com:/home/ec2-user/WWBUILD
-                     scp -o StrictHostKeyChecking=no -i $SSH_KEY .env ec2-user@ec2-13-58-233-86.us-east-2.compute.amazonaws.com:/home/ec2-user/WWBUILD
-                    ssh -o StrictHostKeyChecking=no -i $SSH_KEY ec2-user@ec2-13-58-233-86.us-east-2.compute.amazonaws.com 'node --version'
+                
+                    sh '''
+                        ssh -o StrictHostKeyChecking=no -i $SSH_KEY ec2-user@ec2-13-58-233-86.us-east-2.compute.amazonaws.com 'mkdir WWBUILD'
+                        scp -o StrictHostKeyChecking=no -i $SSH_KEY -r backend ec2-user@ec2-13-58-233-86.us-east-2.compute.amazonaws.com:/home/ec2-user/WWBUILD
+                        scp -o StrictHostKeyChecking=no -i $SSH_KEY .env ec2-user@ec2-13-58-233-86.us-east-2.compute.amazonaws.com:/home/ec2-user/WWBUILD
+                        ssh -o StrictHostKeyChecking=no -i $SSH_KEY ec2-user@ec2-13-58-233-86.us-east-2.compute.amazonaws.com 'node --version'
+                        
+                    '''
+                    } else {
+                        println "we have decided not to go forward with prod build"
+                    }
                     
-                '''
                } 
             }
         }
